Contributors.jsx was in the wrong folder
